+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 2.7.3 - May 18th, 2019
+- Fixed the USACO Training parser
+
+# 2.7.2 - May 13rd, 2019
+- Fixed the USACO Training parser
+
+# 2.7.1 - May 7th, 2019
+- Fixed the HackerEarth parser
+
 # 2.7.0 - April 28th, 2019
 - Added parsers for Baekjoon Online Judge
 - Fixed the CodeChef and the new Google Code Jam parser (thanks [@andrewduffy](https://github.com/andrewduffey) for the Google Code Jam fix)